Separate CML and overall tool run initialization from results_generator
